requirements.txt
appdirs==1.4.4
apturl==0.5.2
beautifulsoup4==4.9.3
blinker==1.4
blis==0.7.4
Brlapi==0.7.0
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
cymem==2.0.5
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
en-core-web-sm==2.3.1
entrypoints==0.3
filelock==3.0.12
gensim==3.8.3
httplib2==0.14.0
idna==2.8
imbalanced-learn==0.7.0
joblib==1.0.0
kazam==1.4.5
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
lyricsgenius==2.0.2
macaroonbakery==1.3.1
matplotlib==3.3.3
murmurhash==1.0.5
netifaces==0.10.4
nltk==3.5
numpy==1.19.4
oauthlib==3.1.0
olefile==0.46
pandas==1.2.0
pbr==5.5.1
pexpect==4.6.0
Pillow==7.0.0
plac==1.1.3
preshed==3.0.5
protobuf==3.6.1
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pylast==4.1.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.15.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
regex==2020.11.13
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
scikit-learn==0.24.0
scipy==1.6.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
smart-open==4.1.0
soupsieve==2.1
spacy==2.3.5
srsly==1.0.5
stevedore==3.3.0
systemd-python==234
thinc==7.4.5
threadpoolctl==2.1.0
tqdm==4.55.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.2.2
virtualenv-clone==0.5.4
wadllib==1.3.3
wasabi==0.8.0
xkit==0.0.0
